+++ CHANGES.md
@@ -1,3 +1,7 @@
+# 0.11.0
+
+* Forward a server context object to all http services. This can be used to lookup peer-socket address and ssl details (if using ssl) for the underlying connection.
+
 # 0.10.1
 
 * Accept a `(string * string) list` as http headers.